Update combat system naming conventions
