+++ Assets/Data/Attacks/TestFireball/TestFireball.asset
@@ -9,10 +9,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3012e45bb62dbea4dabab12a43cb29c7, type: 3}
-  m_Name: TestFireballAttack
+  m_Script: {fileID: 11500000, guid: c7ad955b4b8e17f40ac9aebba29058f6, type: 3}
+  m_Name: TestFireball
   m_EditorClassIdentifier: 
-  actions:
+  effects:
   - {fileID: 11400000, guid: 8fe01f050fb61734ba5d4635b94022b5, type: 2}
   animationController: {fileID: 22100000, guid: 0c9a8c0ef0df9d04980f5a208c7f7af2,
     type: 2}
