+++ Assets/Data/Attacks/TestFireball/TestFireball.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4305619939e4bea4d8cfa6e517c8262f
+guid: eaa6856a482658241a2e5267f2b20a7c
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
