+++ Assets/Scenes/Main.unity
@@ -2206,7 +2206,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   moveSpeed: 3
-  primaryAttack: {fileID: 11400000, guid: 4305619939e4bea4d8cfa6e517c8262f, type: 2}
+  primaryAttack: {fileID: 11400000, guid: eaa6856a482658241a2e5267f2b20a7c, type: 2}
   secondaryAttack: {fileID: 11400000, guid: cb6d74392ec61a147b549286e0221774, type: 2}
 --- !u!95 &2098406990
 Animator:
