+++ Assets/Scripts/Game/Effect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 87be02099ab819746968f6c62b7336bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
