+++ Assets/Scripts/Game/FrontalMelee.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FrontalMelee : MonoBehaviour
+{
+    
+}
