+++ Assets/Scripts/Game/FrontalMelee.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3012e45bb62dbea4dabab12a43cb29c7
+guid: 0c681ddb636ed51468d9b9928a4cab78
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
