+++ Assets/Scripts/Game/PositionalAOE.cs
@@ -2,10 +2,10 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-[CreateAssetMenu(fileName = "New PositionalAOE", menuName = "Actions/PositionalAOE", order = 50)]
-public class PositionalAOE : Action
+[CreateAssetMenu(fileName = "New PositionalAOE", menuName = "Effects/PositionalAOE", order = 50)]
+public class PositionalAOE : Effect
 {
-    [SerializeField] private GameObject effectPrefab;
+    [SerializeField] private GameObject visualPrefab;
     [SerializeField] private int damage = 1;
     [SerializeField] private float range = 1;
     [SerializeField] private float radius = 1;
@@ -13,7 +13,7 @@ public class PositionalAOE : Action
     public override void Invoke()
     {
         Vector2 mousePos = Camera.main.ScreenToWorldPoint(Input.mousePosition);
-        Instantiate(effectPrefab, mousePos, Quaternion.identity);
+        Instantiate(visualPrefab, mousePos, Quaternion.identity);
         Collider2D[] hits = Physics2D.OverlapCircleAll(mousePos, radius);
         foreach (Collider2D hit in hits)
         {
